componentialization
