+++ src/components/food/FoodCalories.jsx
@@ -1,0 +1,10 @@
+export default (nutrientObj) => {
+
+    console.log(nutrientObj);
+
+    return(
+        <div>
+            <p>칼로리: {nutrientObj?.nutrientObj?.calories ?? "-"} kcal</p>
+        </div>
+    );
+}
