+++ src/components/food/FoodManufacture.jsx
@@ -1,0 +1,6 @@
+export default (item) => {
+    return(
+        <div>
+            <p>제조원: {item?.item?.manufacture ?? "-"}</p>
+        </div>
+    );}
