+++ src/components/food/FoodProductGb.jsx
@@ -1,0 +1,7 @@
+export default (item) => {
+    return(
+        <div>
+            <p>식품 종류: {item?.item?.productGb ?? "-"}</p>
+        </div>
+    );
+}
