+++ src/components/food/FoodServing.jsx
@@ -1,0 +1,8 @@
+export default (nutrientObj) => {
+
+    return(
+        <div>
+            <p>1회 제공량: {nutrientObj?.nutrientObj?.serving ?? "-"} g</p>
+        </div>
+    );
+}
